Upgrading PHP code from MySQL to MySQLi.
Note:
The original MySQL extension is now deprecated, and will generate E_DEPRECATED errors (not warnings) when connecting to a database.

Instead, is needed to use the MySQLi or PDO_MySQL extensions.

I decided to use mysqli over PDO_MYSQL for adapting the code fast.

Also, we must take in mind that PDO_MYSQL hasn't got num_rows such as mysqli or mysql.

There is rowCount(), but doesn't work the same way.

See discussions on StackOverFlow:
https://stackoverflow.com/questions/883365/row-count-with-pdo

And PHP.net:
https://php.net/manual/en/pdostatement.rowcount.php#78338

So the code, must be debugged and reanalized before using PDO.